+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/user/AdminUserRepository.java
@@ -1,4 +1,4 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.user;
 
 import com.sloweat.domain.user.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
