+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/user/AdminUserRepositoryCustom.java
@@ -1,7 +1,7 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.user;
 
-import com.sloweat.domain.admin.dto.AdminUserRequest;
-import com.sloweat.domain.admin.dto.AdminUserResponse;
+import com.sloweat.domain.admin.dto.user.AdminUserRequest;
+import com.sloweat.domain.admin.dto.user.AdminUserResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
