+++ backend/sloweat/src/main/java/com/sloweat/domain/admin/repository/user/AdminUserRepositoryCustomImpl.java
@@ -1,11 +1,11 @@
-package com.sloweat.domain.admin.repository;
+package com.sloweat.domain.admin.repository.user;
 
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.core.types.dsl.CaseBuilder;
 import com.querydsl.core.types.dsl.StringExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import com.sloweat.domain.admin.dto.AdminUserRequest;
-import com.sloweat.domain.admin.dto.AdminUserResponse;
+import com.sloweat.domain.admin.dto.user.AdminUserRequest;
+import com.sloweat.domain.admin.dto.user.AdminUserResponse;
 import com.sloweat.domain.admin.dto.QAdminUserResponse;
 import com.sloweat.domain.user.entity.QUser;
 import com.sloweat.domain.user.entity.User.JoinType;
@@ -13,7 +13,6 @@ import java.util.List;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
-import com.sloweat.domain.user.entity.User;
 
 public class AdminUserRepositoryCustomImpl implements AdminUserRepositoryCustom{
   private final JPAQueryFactory queryFactory;
